pyproject.toml
[project]
name = "entomological-label-information-extraction"
version = "1.0.0"
description = "Package for specimen label information extraction and processing"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "GPL-3.0-or-later" }

dependencies = [
        "cer==1.2.0",
        "deskew==1.5.1",
        "detecto==1.2.2",
        "editdistance==0.6.2",
        "Flask==3.1.0",
        "Flask-Cors==4.0.0",
        "gensim==4.3.2",
        "google-cloud-vision==3.7.2",
        "jiwer==3.0.3",
        "kaleido==0.2.1",
        "keras==2.15.0",
        "matplotlib==3.8.4",
        "nltk==3.8.1",
        "numpy==1.26.4",
        "opencv-python==4.9.0.80",
        "pandas==2.2.1",
        "pillow==10.3.0",
        "plotly==5.20.0",
        "plotly-express==0.4.1",
        "pyopenssl",
        "pytesseract==0.3.10",
        "regex==2023.12.25",
        "renku-sphinx-theme==0.4.0",
        "scikit-learn==1.4.1.post1",
        "seaborn==0.13.2",
        "scipy==1.10.1",
        "Sphinx==7.4.7",
        "sphinx-rtd-theme==0.5.1",
        "tensorflow==2.15.0",
        "tesseract==0.1.3",
        "torch==2.2.2",
        "torchvision==0.17.2"
]

[tool.setuptools.packages.find]
where = ["."]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"